+++ src/sympleq/core/phase_correction/__init__.py
@@ -1,0 +1,8 @@
+from .decomposition import clifford_phase_decomposition
+from .solve import pauli_phase_correction, solve_phase_vector_h_from_residual
+
+__all__ = [
+    "clifford_phase_decomposition",
+    "pauli_phase_correction",
+    "solve_phase_vector_h_from_residual",
+]
