+++ src/sympleq/core/phase_correction/decomposition.py
@@ -1,0 +1,3 @@
+from .solve import clifford_phase_decomposition
+
+__all__ = ["clifford_phase_decomposition"]
